// src/storage/types.ts
export enum GBDSuperRegion {
  SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA = "SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA",
  SOUTH_ASIA = "SOUTH_ASIA",
  SUB_SAHARAN_AFRICA = "SUB_SAHARAN_AFRICA",
  LATIN_AMERICA_AND_CARIBBEAN = "LATIN_AMERICA_AND_CARIBBEAN",
  CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA = "CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA",
  NORTH_AFRICA_AND_MIDDLE_EAST = "NORTH_AFRICA_AND_MIDDLE_EAST",
  HIGH_INCOME = "HIGH_INCOME"
}

export enum GBDSubRegion {
  SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA_SUBREGION_SOUTH_EAST_ASIA = "SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA_SUBREGION_SOUTH_EAST_ASIA",
  SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA_SUBREGION_OCEANIA = "SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA_SUBREGION_OCEANIA",
  SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA_SUBREGION_EAST_ASIA = "SOUTH_EAST_ASIA_EAST_ASIA_AND_OCEANIA_SUBREGION_EAST_ASIA",
  SOUTH_ASIA_SUBREGION_SOUTH_ASIA = "SOUTH_ASIA_SUBREGION_SOUTH_ASIA",
  SUB_SAHARAN_AFRICA_SUBREGION_WESTERN = "SUB_SAHARAN_AFRICA_SUBREGION_WESTERN",
  SUB_SAHARAN_AFRICA_SUBREGION_SOUTHERN = "SUB_SAHARAN_AFRICA_SUBREGION_SOUTHERN",
  SUB_SAHARAN_AFRICA_SUBREGION_CENTRAL = "SUB_SAHARAN_AFRICA_SUBREGION_CENTRAL",
  SUB_SAHARAN_AFRICA_SUBREGION_EASTERN = "SUB_SAHARAN_AFRICA_SUBREGION_EASTERN",
  LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_TROPICAL = "LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_TROPICAL",
  LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_CARIBBEAN = "LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_CARIBBEAN",
  LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_ANDEAN = "LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_ANDEAN",
  LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_CENTRAL = "LATIN_AMERICA_AND_CARIBBEAN_SUBREGION_CENTRAL",
  CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA_SUBREGION_CENTRAL_ASIA = "CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA_SUBREGION_CENTRAL_ASIA",
  CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA_SUBREGION_CENTRAL_EUROPE = "CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA_SUBREGION_CENTRAL_EUROPE",
  CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA_SUBREGION_EASTERN_EUROPE = "CENTRAL_EUROPE_EASTERN_EUROPE_AND_CENTRAL_ASIA_SUBREGION_EASTERN_EUROPE",
  NORTH_AFRICA_AND_MIDDLE_EAST_SUBREGION_NORTH_AFRICA_AND_MIDDLE_EAST = "NORTH_AFRICA_AND_MIDDLE_EAST_SUBREGION_NORTH_AFRICA_AND_MIDDLE_EAST",
  HIGH_INCOME_SUBREGION_WESTERN_EUROPE = "HIGH_INCOME_SUBREGION_WESTERN_EUROPE",
  HIGH_INCOME_SUBREGION_SOUTHERN_LATIN_AMERICA = "HIGH_INCOME_SUBREGION_SOUTHERN_LATIN_AMERICA",
  HIGH_INCOME_SUBREGION_NORTH_AMERICA = "HIGH_INCOME_SUBREGION_NORTH_AMERICA",
  HIGH_INCOME_SUBREGION_ASIA_PACIFIC = "HIGH_INCOME_SUBREGION_ASIA_PACIFIC",
  HIGH_INCOME_SUBREGION_AUSTRALASIA = "HIGH_INCOME_SUBREGION_AUSTRALASIA"
}

export enum UNRegion {
  NORTHERN_AFRICA = "NORTHERN_AFRICA",
  EASTERN_AFRICA = "EASTERN_AFRICA",
  MIDDLE_AFRICA = "MIDDLE_AFRICA",
  SOUTHERN_AFRICA = "SOUTHERN_AFRICA",
  WESTERN_AFRICA = "WESTERN_AFRICA",
  CARIBBEAN = "CARIBBEAN",
  CENTRAL_AMERICA = "CENTRAL_AMERICA",
  SOUTH_AMERICA = "SOUTH_AMERICA",
  NORTHERN_AMERICA = "NORTHERN_AMERICA",
  CENTRAL_ASIA = "CENTRAL_ASIA",
  EASTERN_ASIA = "EASTERN_ASIA",
  SOUTH_EASTERN_ASIA = "SOUTH_EASTERN_ASIA",
  SOUTHERN_ASIA = "SOUTHERN_ASIA",
  WESTERN_ASIA = "WESTERN_ASIA",
  EASTERN_EUROPE = "EASTERN_EUROPE",
  NORTHERN_EUROPE = "NORTHERN_EUROPE",
  SOUTHERN_EUROPE = "SOUTHERN_EUROPE",
  WESTERN_EUROPE = "WESTERN_EUROPE",
  AUSTRALIA_AND_NEW_ZEALAND = "AUSTRALIA_AND_NEW_ZEALAND",
  MELANESIA = "MELANESIA",
  MICRONESIA = "MICRONESIA",
  POLYNESIA = "POLYNESIA",
}

export enum WHORegion {
  AFR = "AFR",
  AMR = "AMR",
  SEAR = "SEAR",
  EUR = "EUR",
  EMR = "EMR",
  WPR = "WPR"
}

export interface SarsCov2EstimateDocument {
  _id: ObjectId;
  riskOfBias: string | undefined;
  antibodies: string[];
  isotypes: string[];
  isWHOUnityAligned: boolean;
  testType: string[];
  ageGroup: string | undefined;
  sex: string | undefined;
  studyName: string;
  sourceType: string | undefined;
  populationGroup: string | undefined;
  latitude: number;
  longitude: number;
  whoRegion: WHORegion | undefined;
  unRegion: UNRegion | undefined;
  gbdSuperRegion: GBDSuperRegion | undefined;
  gbdSubRegion: GBDSubRegion | undefined;
  country: string;
  countryAlphaTwoCode: string;
  countryAlphaThreeCode: string;
  state: string | undefined;
  county: string | undefined;
  city: string | undefined;
  scope: string | undefined;
  samplingEndDate: Date | undefined;
  samplingStartDate: Date | undefined;
  samplingMidDate: Date | undefined;
  publicationDate: Date | undefined;
  countryPeopleVaccinatedPerHundred: number | undefined;
  countryPeopleFullyVaccinatedPerHundred: number | undefined;
  countryPositiveCasesPerMillionPeople: number | undefined;
  countryPopulation: number | undefined;
  denominatorValue: number | undefined;
  numeratorValue: number | undefined;
  seroprevalence: number | undefined;
  estimateName: string | undefined;
  partitionKey: number;
  url: string | undefined;
  createdAt: Date;
  updatedAt: Date;
}